Local AI Agent Tutorial/requirements.txt
langchain
langchain-ollama
langchain-chroma
python-dotenv
pandas